Add a sample query result to the "fixtures" directory
This one was taken straight from the Star Wars example in the Relay
repo. It's a tiny result, and one that doesn't really exhibit any
redundancy that we could usefully exploit via a JSON Graph like
representation of it, but it will serve for the purposes of testing out
some initial ideas.