Increase result size in todo example
I want a bigger, more meaningful query with some redundancy in it for
the purposes of testing my hackathon stuff. So, hack the TodoMVC example
to return a much larger result set (about 49,484 bytes) for its initial
query.

Note: not intended for shipping. This can be rolled back later on.