Add "arrayrefize" strategy
Here, instead of storing lenghty keys whenever we want to make a pointer
into the normalized `__nodes__` structure, we store a simple index into
an array.

This is smaller than the `thin` strategy (which is itself smaller than
the `flatten` strategy), but still doesn't beat plain old gzip on the
denormalized result object, at least for this sample fixture.