Add some decoding timings
I tried to use the zlibjs library here to reverse the gzipping, but it
proved too buggy and unreliable (it would either succeed, throw an
error, or crash the process, depending on the example input...)

I'll add more decoders later.

Note that the Leon format is supposedly like FlatBuffers in that you can
read out of it without deserializing, but I'll worry about that later.